+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- API: Implement error reporting using Sentry.
+- Frontend: Implement error reporting using Sentry.
+
 ## [0.1.0] - 2018-10-19
 
 This is the initial release of the Appsemble platform.